REFAC: modified file structure and optimize codebase for frontend
