+++ src/types/index.ts
@@ -3,3 +3,4 @@ export * from "./progress";
 export * from "./annotation";
 export * from "./plan";
 export * from "./collection";
+export * from "./context";
