+++ src/types/progress.ts
@@ -1,17 +1,12 @@
-export interface ProgressInfo {
-  current_position?: string | null;
-  progress_percentage?: number;
-}
-
-export interface ProgressItem {
+export interface ReadingProgress {
+  id: number;
   book_id: number;
+  current_position: string;
+  chapter_title: string | null;
+  page_number: number | null;
   progress_percentage: number;
   last_read_at: string | null;
 }
 
-export interface ReadingProgress {
-  progress_percentage?: number;
-  chapter_title?: string | null;
-  page_number?: number | null;
-  last_read_at?: string | null;
-}
+export type ProgressItem = ReadingProgress;
+export type ProgressInfo = Partial<ReadingProgress>;
